+++ NEWS.md
@@ -4,6 +4,7 @@ Changelog
 Version 0.4.1
 -------------
 
+* Added an option to generate palette colors from the currently open image.
 * Added unit tests to ensure the recoloring algorithm never breaks silently
   before a release.